[COMMON] media: mfc: support 64bit system
Conflicts:
	drivers/media/platform/exynos/mfc/s5p_mfc.c

+++ drivers/media/platform/exynos/mfc/s5p_mfc_common.h
@@ -250,7 +250,7 @@ struct s5p_mfc_buf_size_v6 {
 };
 
 struct s5p_mfc_buf_size {
-	unsigned int firmware_code;
+	size_t firmware_code;
 	unsigned int cpb_buf;
 	void *buf;
 };
